Moved classes from f_hot-file to their own files
